+++ internal/collector/collector.go
@@ -147,23 +147,23 @@ func (c *Collector) Collect(ch chan<- prometheus.Metric) {
 
 		ch <- prometheus.MustNewConstMetric(c.processCPU, prometheus.GaugeValue, stats.Process.Cpu)
 
-		ch <- prometheus.MustNewConstMetric(c.processMemoryBytes, prometheus.GaugeValue, stats.Process.MemoryBytes)
-		ch <- prometheus.MustNewConstMetric(c.diskIoReadBytes, prometheus.GaugeValue, stats.DiskIo.ReadBytes)
-		ch <- prometheus.MustNewConstMetric(c.diskIoWrittenBytes, prometheus.GaugeValue, stats.DiskIo.WrittenBytes)
-		ch <- prometheus.MustNewConstMetric(c.diskIoReadOps, prometheus.GaugeValue, stats.DiskIo.ReadOps)
-		ch <- prometheus.MustNewConstMetric(c.diskIoWriteOps, prometheus.GaugeValue, stats.DiskIo.WriteOps)
-		ch <- prometheus.MustNewConstMetric(c.tcpSentBytes, prometheus.GaugeValue, stats.Tcp.SentBytes)
-		ch <- prometheus.MustNewConstMetric(c.tcpReceivedBytes, prometheus.GaugeValue, stats.Tcp.ReceivedBytes)
-		ch <- prometheus.MustNewConstMetric(c.tcpConnections, prometheus.GaugeValue, stats.Tcp.Connections)
+		ch <- prometheus.MustNewConstMetric(c.processMemoryBytes, prometheus.GaugeValue, float64(stats.Process.MemoryBytes))
+		ch <- prometheus.MustNewConstMetric(c.diskIoReadBytes, prometheus.GaugeValue, float64(stats.DiskIo.ReadBytes))
+		ch <- prometheus.MustNewConstMetric(c.diskIoWrittenBytes, prometheus.GaugeValue, float64(stats.DiskIo.WrittenBytes))
+		ch <- prometheus.MustNewConstMetric(c.diskIoReadOps, prometheus.GaugeValue, float64(stats.DiskIo.ReadOps))
+		ch <- prometheus.MustNewConstMetric(c.diskIoWriteOps, prometheus.GaugeValue, float64(stats.DiskIo.WriteOps))
+		ch <- prometheus.MustNewConstMetric(c.tcpSentBytes, prometheus.GaugeValue, float64(stats.Tcp.SentBytes))
+		ch <- prometheus.MustNewConstMetric(c.tcpReceivedBytes, prometheus.GaugeValue, float64(stats.Tcp.ReceivedBytes))
+		ch <- prometheus.MustNewConstMetric(c.tcpConnections, prometheus.GaugeValue, float64(stats.Tcp.Connections))
 
 		for _, queue := range stats.Queues {
-			ch <- prometheus.MustNewConstMetric(c.queueLength, prometheus.GaugeValue, queue.Length, queue.Name)
-			ch <- prometheus.MustNewConstMetric(c.queueItemsProcessed, prometheus.CounterValue, queue.ItemsProcessed, queue.Name)
+			ch <- prometheus.MustNewConstMetric(c.queueLength, prometheus.GaugeValue, float64(queue.Length), queue.Name)
+			ch <- prometheus.MustNewConstMetric(c.queueItemsProcessed, prometheus.CounterValue, float64(queue.ItemsProcessed), queue.Name)
 		}
 
 		for _, drive := range stats.Drives {
-			ch <- prometheus.MustNewConstMetric(c.driveTotalBytes, prometheus.GaugeValue, drive.TotalBytes, drive.Name)
-			ch <- prometheus.MustNewConstMetric(c.driveAvailableBytes, prometheus.GaugeValue, drive.AvailableBytes, drive.Name)
+			ch <- prometheus.MustNewConstMetric(c.driveTotalBytes, prometheus.GaugeValue, float64(drive.TotalBytes), drive.Name)
+			ch <- prometheus.MustNewConstMetric(c.driveAvailableBytes, prometheus.GaugeValue, float64(drive.AvailableBytes), drive.Name)
 		}
 
 		for _, projection := range stats.Projections {
@@ -173,16 +173,16 @@ func (c *Collector) Collect(ch chan<- prometheus.Metric) {
 			}
 			ch <- prometheus.MustNewConstMetric(c.projectionRunning, prometheus.GaugeValue, running, projection.Name)
 			ch <- prometheus.MustNewConstMetric(c.projectionProgress, prometheus.GaugeValue, projection.Progress, projection.Name)
-			ch <- prometheus.MustNewConstMetric(c.projectionEventsProcessedAfterRestart, prometheus.CounterValue, projection.EventsProcessedAfterRestart, projection.Name)
+			ch <- prometheus.MustNewConstMetric(c.projectionEventsProcessedAfterRestart, prometheus.CounterValue, float64(projection.EventsProcessedAfterRestart), projection.Name)
 		}
 
 		for _, subscription := range stats.Subscriptions {
-			ch <- prometheus.MustNewConstMetric(c.subscriptionTotalItemsProcessed, prometheus.CounterValue, subscription.TotalItemsProcessed, subscription.EventStreamID, subscription.GroupName)
-			ch <- prometheus.MustNewConstMetric(c.subscriptionLastProcessedEventNumber, prometheus.GaugeValue, subscription.LastProcessedEventNumber, subscription.EventStreamID, subscription.GroupName)
-			ch <- prometheus.MustNewConstMetric(c.subscriptionLastKnownEventNumber, prometheus.GaugeValue, subscription.LastKnownEventNumber, subscription.EventStreamID, subscription.GroupName)
-			ch <- prometheus.MustNewConstMetric(c.subscriptionConnectionCount, prometheus.GaugeValue, subscription.ConnectionCount, subscription.EventStreamID, subscription.GroupName)
-			ch <- prometheus.MustNewConstMetric(c.subscriptionTotalInFlightMessages, prometheus.GaugeValue, subscription.TotalInFlightMessages, subscription.EventStreamID, subscription.GroupName)
-			ch <- prometheus.MustNewConstMetric(c.subscriptionTotalNumberOfParkedMessages, prometheus.GaugeValue, subscription.TotalNumberOfParkedMessages, subscription.EventStreamID, subscription.GroupName)
+			ch <- prometheus.MustNewConstMetric(c.subscriptionTotalItemsProcessed, prometheus.CounterValue, float64(subscription.TotalItemsProcessed), subscription.EventStreamID, subscription.GroupName)
+			ch <- prometheus.MustNewConstMetric(c.subscriptionLastProcessedEventNumber, prometheus.GaugeValue, float64(subscription.LastProcessedEventNumber), subscription.EventStreamID, subscription.GroupName)
+			ch <- prometheus.MustNewConstMetric(c.subscriptionLastKnownEventNumber, prometheus.GaugeValue, float64(subscription.LastKnownEventNumber), subscription.EventStreamID, subscription.GroupName)
+			ch <- prometheus.MustNewConstMetric(c.subscriptionConnectionCount, prometheus.GaugeValue, float64(subscription.ConnectionCount), subscription.EventStreamID, subscription.GroupName)
+			ch <- prometheus.MustNewConstMetric(c.subscriptionTotalInFlightMessages, prometheus.GaugeValue, float64(subscription.TotalInFlightMessages), subscription.EventStreamID, subscription.GroupName)
+			ch <- prometheus.MustNewConstMetric(c.subscriptionTotalNumberOfParkedMessages, prometheus.GaugeValue, float64(subscription.TotalNumberOfParkedMessages), subscription.EventStreamID, subscription.GroupName)
 			ch <- prometheus.MustNewConstMetric(c.subscriptionOldestParkedMessage, prometheus.GaugeValue, subscription.OldestParkedMessageAgeInSeconds, subscription.EventStreamID, subscription.GroupName)
 		}
